extract Collapsible component from Workstation
